update some styling (#23)
Update the padding around the command inputs to improve visual grouping
with the next command output rather than previous comamnd output.
Additionally, remove the bolding of the data-dt as the colour scheme
makes it unnecessary, also prevents clashing with the input. Finally,
change the command input text to a brighter white for improved contrast
with the background.